# api/http/openapi.yaml
openapi: "3.0.3"
info:
  title: Breakfix API
  version: "1.0.0"
  description: SRE/DevOps interview practice platform
servers:
  - url: http://localhost:9090/api
    description: Local dev
paths:
...
  /authoring/sessions:
    post:
      operationId: createAuthoringSession
      summary: Create a human-reviewed challenge authoring session
      tags: [authoring]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Authoring session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthoringSession"
  /authoring/sessions/current:
    get:
      operationId: getCurrentAuthoringSession
      summary: Resume the latest unfinished authoring session for the current user
      tags: [authoring]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current unfinished authoring session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthoringSession"
        "404":
          $ref: "#/components/responses/Error"
  /authoring/sessions/{id}:
    get:
      operationId: getAuthoringSession
      summary: Read the current authoring revision, conversation, assets, and diff
      tags: [authoring]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthoringSessionID"
      responses:
        "200":
          description: Current authoring session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthoringSession"
        "404":
          $ref: "#/components/responses/Error"
  /authoring/sessions/{id}/messages:
    post:
      operationId: sendAuthoringMessage
      summary: Send one natural-language instruction to the authoring agent
      tags: [authoring]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthoringSessionID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthoringMessageRequest"
      responses:
        "200":
          description: Server-sent authoring ready, delta, complete, and error events
          content:
            text/event-stream:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/Error"
  /authoring/sessions/{id}/generate:
    post:
      operationId: confirmAuthoringGeneration
      summary: Explicitly start generation and real verification for the current intent revision
      tags: [authoring]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthoringSessionID"
      responses:
        "200":
          description: Generator Run and verification started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthoringSession"
  /authoring/sessions/{id}/publish:
    post:
      operationId: publishAuthoringRevision
      summary: Explicitly publish the current verified authoring revision
      tags: [authoring]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthoringSessionID"
      responses:
        "200":
          description: Verified revision published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthoringSession"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  parameters:
    AuthoringSessionID:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    AuthoringMetadata:
      type: object
      required:
        - title
        - difficulty
        - description
        - runtime
      properties:
        title:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        description:
          type: string
        runtime:
          type: string
          enum: [node, k8s]
    AuthoringCheckpoint:
      type: object
      required: [id, title, markdown, position]
      properties:
        id:
          type: string
        title:
          type: string
        markdown:
          type: string
        position:
          type: integer
    AuthoringPlan:
      type: object
      required: [metadata, overview, checkpoints]
      properties:
        metadata:
          $ref: "#/components/schemas/AuthoringMetadata"
        overview:
          type: string
        checkpoints:
          type: array
          items:
            $ref: "#/components/schemas/AuthoringCheckpoint"
    VerifiedCheckpoint:
      type: object
      required: [id, title, description]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        hint:
          type: string
        node:
          type: string
    VerifiedChallenge:
      type: object
      required: [metadata, checkpoints]
      properties:
        metadata:
          $ref: "#/components/schemas/AuthoringMetadata"
        checkpoints:
          type: array
          items:
            $ref: "#/components/schemas/VerifiedCheckpoint"
    AuthoringCandidate:
      type: object
      required: [id, generator_run_id, archive_sha256]
      properties:
        id:
          type: string
        generator_run_id:
          type: string
        archive_sha256:
          type: string
    AuthoringGenerationWorkflow:
      type: object
      required: [id, state, state_attempt, created_at, updated_at]
      properties:
        id:
          type: string
        state:
          type: string
          enum: [Queued, Generating, Judging, Building, ArtifactPublishing, Verifying, NeedsAuthorReview, ChallengePublishing, CleaningUp, Completed, Failed, Cancelled]
        state_attempt:
          type: integer
          minimum: 0
        candidate_revision_id:
          type: string
        deadline_at:
          type: string
          format: date-time
          nullable: true
        last_error:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AuthoringVerificationReport:
      type: object
      required: [passed, answers, checkpoints, summary]
      properties:
        passed:
          type: boolean
        summary:
          type: string
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AuthoringExecutionResult"
        checkpoints:
          type: array
          items:
            $ref: "#/components/schemas/AuthoringCheckpointResult"
    AuthoringExecutionResult:
      type: object
      required: [location, exit_code]
      properties:
        location:
          type: string
        exit_code:
          type: integer
        stdout:
          type: string
        stderr:
          type: string
    AuthoringCheckpointResult:
      type: object
      required: [id, passed, summary]
      properties:
        id:
          type: string
        passed:
          type: boolean
        summary:
          type: string
        details:
          type: string
    AuthoringChange:
      type: object
      required: [kind, summary, difficulty_impact, revision]
      properties:
        kind:
          type: string
        summary:
          type: string
        difficulty_impact:
          type: string
        revision:
          type: integer
    AuthoringMessage:
      type: object
      required: [id, role, content, created_at]
      properties:
        id:
          type: string
        role:
          type: string
          enum: [user, agent, system, event]
        content:
          type: string
        changes:
          type: array
          items:
            $ref: "#/components/schemas/AuthoringChange"
        created_at:
          type: string
          format: date-time
    AuthoringAsset:
      type: object
      required: [path, content]
      properties:
        path:
          type: string
        content:
          type: string
    AuthoringFileDiff:
      type: object
      required: [path, diff]
      properties:
        path:
          type: string
        diff:
          type: string
    AuthoringMessageRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
    AuthoringSession:
      type: object
      required: [id, state, authoring_turn_active, intent_revision, visible_revision, updated_at, intent, messages, assets, diff]
      properties:
        id:
          type: string
        state:
          type: string
          enum: [DraftConversation, IntentReview, Published]
        authoring_turn_active:
          type: boolean
          description: Whether the durable authoring Agent Run is pending or running. While true, the plan is being updated privately and author actions are unavailable.
        intent_revision:
          type: integer
        visible_revision:
          type: integer
        publish_challenge_id:
          type: string
        workflow:
          $ref: "#/components/schemas/AuthoringGenerationWorkflow"
        last_error:
          type: string
        updated_at:
          type: string
          format: date-time
        intent:
          $ref: "#/components/schemas/AuthoringPlan"
        candidate:
          $ref: "#/components/schemas/AuthoringCandidate"
        verified:
          $ref: "#/components/schemas/VerifiedChallenge"
        verification:
          $ref: "#/components/schemas/AuthoringVerificationReport"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AuthoringMessage"
        assets:
          type: array
          items:
            $ref: "#/components/schemas/AuthoringAsset"
        diff:
          type: array
          items:
            $ref: "#/components/schemas/AuthoringFileDiff"